+++ lib/puppet/type/wls_coherence_cluster.rb
@@ -42,6 +42,7 @@ module Puppet
     parameter :name
     parameter :timeout
     parameter :coherence_cluster_name
+    parameter :storage_enabled
     property :clusteringmode
     property :unicastaddress
     property :unicastport