+++ lib/puppet/type/wls_coherence_cluster/storage_enabled.rb
@@ -1,0 +1,10 @@
+newparam(:storage_enabled) do
+  include EasyType
+  include EasyType::Validators::Name
+
+  desc 'Enables coherence storage on its cluster/server targets'
+  newvalues(1, 0)
+
+  defaultto '0'
+
+end