+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.22
 - fix for copydomain when the standard domains location is used
+- enable storage on wls_coherence_cluster
 
 ## 1.0.21
 - puppet custom type for 12.1.2, 12.1.3 RCU, checks first if it already exists